fast-style-transfer/flask_app_wip/requirements.txt
Flask==1.0.2
matplotlib==2.2.3
numpy==1.15.1
Pillow==5.2.0
torch==1.0.1.post2
torchvision==0.2.1  